requirements.txt
bcrypt==3.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
DateTime==5.2
Deprecated==1.2.14
docopt==0.6.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Limiter==3.5.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.2.0
ordered-set==4.1.0
packaging==23.1
pipreqs==0.4.13
protobuf==4.24.3
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.6.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==2.3.7
wrapt==1.15.0
yarg==0.1.9
zope.interface==6.0
python-telegram-bot
flask-swagger-ui
py-vapid
cryptography
pywebpush
